% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapxplore-package.R
\docType{package}
\name{gapxplore-package}
\alias{gapxplore}
\alias{gapxplore-package}
\title{gapxplore: Interactive Gapminder Explorer}
\description{
Tools for packaging a tidy Gapminder dataset and a Shiny app
for ETC5523 Assignment 4 (data, app launcher, and docs).
}
\author{
\strong{Maintainer}: Midhun Unnikrishnan \email{munn0002@student.monash.edu}

}
\keyword{internal}
